Mastermind Alpha Update
Hey guys! Thanks for your patience with me in developing this alpha of Mastermind. I have removed the One Time Pad software and will add it later when the alpha is complete. I left the actual alpha in the group directory and current iterative code in my folder. Ignore my folder as it is for me to work with since I have two main workstations and git makes it easy to update changes. Anyways enjoy playing with the executable it works on windows! Let me know of any necessary additional features or feedback.